+++ rusqlite_migration/src/lib.rs
@@ -420,14 +420,6 @@ impl<'m> Migrations<'m> {
         Ok(Self { ms: migrations })
     }
 
-    /// **Deprecated**: [`Migrations`] now implements [`FromIterator`], so use [`Migrations::from_iter`] instead.
-    ///
-    /// Performs allocations transparently.
-    #[deprecated = "Use the `FromIterator` trait implementation instead. For instance, you can call Migrations::from_iter."]
-    pub fn new_iter<I: IntoIterator<Item = M<'m>>>(ms: I) -> Self {
-        Self::new(Vec::from_iter(ms))
-    }
-
     fn db_version_to_schema(&self, db_version: usize) -> SchemaVersion {
         match db_version {
             0 => SchemaVersion::NoneSet,